Simplify USB ENDPTCTRL registers in PAC
The ENDPTCTRL registers for endpoints 1 through 7
are the same. The only one that's unique is
endpoint 0, which can only be a control endpoint.

We add a few dim tags to the SVD file to remove
the redundant register definitions. This simplifies
the Rust implementation; now there's only two types
of ENDPTCTRL registers to work with, instead of 8.