Shim the cortex-m-rt crate for RTIC
The RTIC crate directly depends on the typical cortex-m-rt runtime
crate. This crate doesn't yet support the start-up needs for the
Teensy 4. So, if we want to use RTIC on the Teensy 4, we need to
trick the RTIC crate.

The commit renames the teensy4-rt crate to cortex-m-rt. Obviously,
this won't work at scale, since we can't use this name! But, we can
use a [patch] directive to trick RTIC into thinking our runtime crate
is cortex-m-rt. the teensy4-rt crate is API compatible, so it's a
drop-in.

This commit will never be released; it's for a POC.